from-scratch/neural-uplink/requirements.txt
aiohttp==3.9.1
pyyaml==6.0.1
psutil==5.9.8
chroma-db==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.4
python-docx==1.1.0